+++ InscryptionCommunityPatch/ResourceManagers/ActOneEnergyDrone.cs
@@ -188,7 +188,7 @@ public static class EnergyDrone
             baseResourceManager.PlayerMaxEnergy = 0;
         }
 
-        if (EnergyConfig.ConfigDrone)
+        if (EnergyConfig.ConfigDrone && ResourceDrone.m_Instance != null)
         {
             ResourceDrone.Instance.CloseAllCells(false);
             ResourceDrone.Instance.SetOnBoard(false, false);
@@ -205,11 +205,11 @@ public static class EnergyDrone
     {
         if (__instance is Part1ResourcesManager && EnergyConfig.ConfigDrone)
         {
-            ResourceDrone.Instance.SetOnBoard(true, false);
+            ResourceDrone.Instance?.SetOnBoard(true, false);
             if (EnergyConfig.ConfigDroneMox)
             {
                 PatchPlugin.Logger.LogDebug("Setting up extra resources for the drone.");
-                ResourceDrone.Instance.Gems.SetAllGemsOn(false, true);
+                ResourceDrone.Instance?.Gems.SetAllGemsOn(false, true);
             }
         }
     }
@@ -221,7 +221,7 @@ public static class EnergyDrone
         if (__instance is Part1ResourcesManager && EnergyConfig.ConfigDrone)
         {
             int cellsToOpen = __instance.PlayerMaxEnergy - 1;
-            ResourceDrone.Instance.OpenCell(cellsToOpen);
+            ResourceDrone.Instance?.OpenCell(cellsToOpen);
             yield return new WaitForSeconds(0.4f);
         }
 
@@ -237,7 +237,7 @@ public static class EnergyDrone
             int num;
             for (int i = __instance.PlayerEnergy - amount; i < __instance.PlayerEnergy; i = num + 1)
             {
-                ResourceDrone.Instance.SetCellOn(i, true, false);
+                ResourceDrone.Instance?.SetCellOn(i, true, false);
                 yield return new WaitForSeconds(0.05f);
                 num = i;
             }
@@ -259,7 +259,7 @@ public static class EnergyDrone
                     __instance.transform.position, 0.4f, 0f,
                     new AudioParams.Pitch(0.9f + (float)(__instance.PlayerEnergy + i) * 0.05f), null, null, null,
                     false);
-                ResourceDrone.Instance.SetCellOn(i, false, false);
+                ResourceDrone.Instance?.SetCellOn(i, false, false);
                 yield return new WaitForSeconds(0.05f);
                 num = i;
             }
@@ -299,7 +299,7 @@ public static class EnergyDrone
     private static void ResourcesManager_SetGemOnImmediate(GemType gem, bool on, ResourcesManager __instance)
     {
         if (__instance is Part1ResourcesManager)
-            ResourceDrone.Instance.Gems.SetGemOn(gem, on, false);
+            ResourceDrone.Instance.Gems?.SetGemOn(gem, on, false);
     }
 
     [HarmonyPatch(typeof(TurnManager), nameof(TurnManager.DoUpkeepPhase))]